+++ JQ_CPP/JQ_Pluralsight_Code/MacrosEtc.cpp
