+++ JQ_CPP/JQ_Pluralsight_Code/Resource.h
