+++ JQ_CPP/JQ_Pluralsight_Code/Strings.cpp
