+++ JQ_CS/JQ_Working_Code/HelloWorld.cs
@@ -1,0 +1,17 @@
+// Hello World - Test C# program
+
+// A skeleton of a C# program 
+using System;
+namespace YourNamespace
+{
+    class YourMainClass
+    {
+        static void Main(string[] args) 
+        {
+		//Your program starts here...
+		Console.WriteLine("Hello World");
+		
+		return 0;
+        }
+    }
+}
