--- SkeletonProgram.cs
@@ -1,37 +1,0 @@
-// A skeleton of a C# program 
-using System;
-namespace YourNamespace
-{
-    class YourClass
-    {
-    }
-
-    struct YourStruct
-    {
-    }
-
-    interface IYourInterface 
-    {
-    }
-
-    delegate int YourDelegate();
-
-    enum YourEnum 
-    {
-    }
-
-    namespace YourNestedNamespace
-    {
-        struct YourStruct 
-        {
-        }
-    }
-
-    class YourMainClass
-    {
-        static void Main(string[] args) 
-        {
-            //Your program starts here...
-        }
-    }
-}
